Adicionando fridge game
