+++ core/src/br/cefetmg/games/minigames/TheFridgeGame.java
@@ -1,0 +1,67 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package br.cefetmg.games.minigames;
+import br.cefetmg.games.Config;
+import br.cefetmg.games.minigames.util.MiniGameState;
+import br.cefetmg.games.minigames.util.MiniGameStateObserver;
+import br.cefetmg.games.minigames.util.TimeoutBehavior;
+import br.cefetmg.games.screens.BaseScreen;
+import com.badlogic.gdx.graphics.Texture;
+
+/**
+ *
+ * @author sarit
+ */
+public class TheFridgeGame extends MiniGame {
+    
+    private Texture texturaFundo;
+   
+
+    public TheFridgeGame(BaseScreen screen, GameStateObserver observer, float difficulty) {
+        super(screen, difficulty, 10000,
+                TimeoutBehavior.FAILS_WHEN_MINIGAME_ENDS, observer);
+    }
+
+    @Override
+    protected void onStart() {
+        this.texturaFundo = screen.assets.get(
+               "super-micro-jogo/personagem.png", Texture.class);
+        
+        // ...
+    }
+    // ...
+
+    @Override
+    protected void configureDifficultyParameters(float difficulty) {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public void onHandlePlayingInput() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public void onUpdate(float dt) {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public void onDrawGame() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public String getInstructions() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public boolean shouldHideMousePointer() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+}
+}
